move generateOffer func to separate file
